Break up both the bbcode and the xhtml plugins into ES modules
#733
#584—but not really since globals are still used. A fully modular component based setup requires string based options to go, and since that is a BC break, it should target v4.0.0.
Also another step in making #402 a reality
