+++ .eslintrc.json
@@ -30,10 +30,12 @@
     "no-undef": "error",
     "no-unused-vars": "error",
     "strict": "error",
-    "max-params": ["error", 4],
-    "max-depth": ["error", 4],
+    "max-params": ["warn", 4],
+    "max-depth": ["warn", 4],
     "max-len": ["error", {
       "code": 80,
+      "ignoreStrings": true,
+      "ignoreTemplateLiterals": true,
       "ignoreUrls": true,
       "ignoreRegExpLiterals": true
     }],
