--- src/formats/.eslintrc.json
@@ -1,5 +1,0 @@
-{
-  "parserOptions": {
-    "sourceType": "script"
-  }
-}
